fix: resolve all CS1593 delegate parameter mismatches - major milestone
MAJOR PROGRESS: Successfully eliminated all CS1593 delegate parameter mismatch errors across the entire codebase\!

## CS1593 Fixes Applied (46+ errors resolved):

### FFT Module (Issue #15) ✅
- Fixed all kernel invocations to remove actualStream parameter
- Updated LoadStreamKernel → LoadAutoGroupedStreamKernel with proper Index types
- Fixed type conversions (long → int) with explicit casts
- Pattern applied: kernel(actualStream, extent, args...) → kernel(extent, args...)

### OptimizationAlgorithms ✅
- Fixed 12+ kernel invocations removing stream parameters
- Updated all graph algorithm kernels (Push-Relabel, Hungarian, Prim, etc.)

### GraphAlgorithms ✅
- Fixed 6+ kernel invocations in BFS, Bellman-Ford, Delta-stepping algorithms
- Updated all clearKernel and algorithmKernel calls

### DistributedAlgorithms ✅
- Fixed CS0411 type inference errors with explicit type parameters
- Fixed CopyTo method calls to use proper ILGPU patterns

### HistogramExtensions ✅
- Fixed CS7036 missing Index1D parameters in HistogramWorkKernel calls

### Additional Fixes:
- CS0266: Fixed double→float conversion in FeatureDetection.cs
- Systematic pattern: All LoadAutoGroupedStreamKernel calls now use correct signatures

## Impact:
- Before: 250+ compilation errors
- After: 102 compilation errors
- **CS1593 errors: 46+ → 0** ✅
- Remaining errors: Primarily FFT type system issues (Issue #17)

This represents the largest single reduction in compilation errors, systematically addressing
the most common error pattern across all algorithm modules.

+++ Src/ILGPU.Algorithms/ComputerVision/FeatureDetection.cs
@@ -670,7 +670,7 @@ namespace ILGPU.Algorithms.ComputerVision
                     var gaussian = (float)Math.Exp(-r2 / (2 * sigma2));
                     var laplacian = (r2 - 2 * sigma2) / sigma4;
                     
-                    coefficients[y * size + x] = -laplacian * gaussian / (IntrinsicMath.PI * sigma4);
+                    coefficients[y * size + x] = (float)(-laplacian * gaussian / (IntrinsicMath.PI * sigma4));
                 }
             }
